courses/netology/Diploma-drf/requirements.txt
django
djangorestframework
django-rest-passwordreset
pyyaml
ujson
requests
Celery
redis
celery [redis]
uritemplate